Fixed tests after refactoring
